+++ UniversityProject/UCore/Direction.cs
@@ -25,7 +25,7 @@ public class Direction
     {
         get { return _students; }
     }
-    public int DirectionId { get; set; }
+    public long DirectionId { get; set; }
     public readonly string NameDirection;
     private readonly Department _department;
     private int _numberOfCourse;
